# docs/reference/openapi.yaml
openapi: 3.1.0
...
info:
  title: switchboard HTTP API
  version: 0.1.0
  description: |
    Inbound webhook ingestion and the local-only web UI for **switchboard**.

    ## Three groups of endpoints
    - **Webhook ingestion** (`/webhooks/w/{token}`): producers POST to the ingest URL an endpoint's
      `create_webhook` returned. Signed source types are verified-or-rejected (401, not persisted);
      a `generic` webhook is authenticated by the unguessable token in its URL.
    - **Operator API** (`/api/v1/*`): OAuth-guarded (operator grants) — register agents, vend
      endpoints, list what you own. The `switchboard` CLI is its reference client.
    - **Web UI** (`/`, `/log`, `/settings`) + the SSE stream (`/events`):
      server-rendered HTML (html/template + HTMX) plus a `text/event-stream` for live updates.

    ## Authentication posture (ADR-0001, brief §8)
    The service is **localhost-bound by default and ships no in-app auth for the UI**. If it is
    ever exposed on the homelab LAN it MUST sit behind Caddy `forward_auth` — auth is provided by
    the reverse proxy, not built into this app for the MVP. The UI/SSE/health endpoints below are
    therefore marked `security: []` (no in-app auth) **by design**, not by oversight.

    For **webhook** endpoints, "authentication" means per-provider signature verification
    (`signed` providers) or a shared-secret token that authenticates the caller (`generic`). See ADR-0003.
  license:
    name: MIT
...
  contact:
...
servers:
...
paths:
...
  /api/v1/webhooks:
    get:
      tags: [operator-api]
      operationId: listWebhooks
      summary: List the webhooks your endpoints own (never ingest URLs or secrets)
      description: |
        Every webhook whose owning endpoint belongs to the authenticated human, active endpoints
        first, a revoked endpoint's included (SPEC-0035). Ingest URLs and signing secrets are never
        returned: they are revealed once, to the owning endpoint, at create or rotate.
      security:
        - operatorOAuth: []
      responses:
        '200':
          description: The webhooks in the caller's reach.
          content:
            application/json:
              schema:
                type: object
                required: [webhooks]
                properties:
                  webhooks:
                    type: array
                    items: { $ref: '#/components/schemas/WebhookListOut' }
        '401':
          description: Missing, expired, revoked, or non-operator bearer (an sbk_ endpoint credential included).
        '429':
          description: The caller's per-human read budget is spent; see Retry-After.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
  /api/v1/webhooks/{webhook_id}/rules:
    parameters:
      - name: webhook_id
        in: path
        required: true
        description: A webhook whose owning endpoint belongs to the caller (see listWebhooks).
        schema: { type: string, format: uuid }
    get:
      tags: [operator-api]
      operationId: getWebhookRules
      summary: Read a webhook's routing rules, default, params and grant
      description: |
        The shape `list_webhook_rules` returns (ADR-0024, SPEC-0020), from the same code. The
        response is also a valid body for putWebhookRules.
      security:
        - operatorOAuth: []
      responses:
        '200':
          description: The webhook's routing configuration and what its actions may reach.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookRules' }
        '401':
          description: Missing, expired, revoked, or non-operator bearer.
        '404':
          description: No such webhook in the caller's reach (another human's and an unknown id are the same answer).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
    put:
      tags: [operator-api]
      operationId: putWebhookRules
      summary: Replace a webhook's routing configuration
      description: |
        `set_webhook_rules` through the same save path: validation against the grant, a dry run
        against the webhook's 50 latest deliveries, and a row-locked write that answers conflict if
        the stored rules changed meanwhile. Any failure keeps the previous configuration. A body
        without a `params` key KEEPS the stored params; `"params": null` clears them (the MCP
        `set_webhook_rules` now gives an omitted `params` the same keep answer, and clears only on
        an explicit `"params": {}`). Read-only fields of a GET response
        are ignored.
      security:
        - operatorOAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WebhookRulesIn' }
      responses:
        '200':
          description: Saved; the stored configuration.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookRules' }
        '400':
          description: |
            The body does not decode or has no `rules`, a rule is invalid (`invalid_rule`,
            `invalid_expression`, `forbidden_function`, `too_many_rules`, `invalid_params`), or a rule
            faults on a recent delivery (`invalid_argument`, naming the rule and the events).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '401':
          description: Missing, expired, revoked, or non-operator bearer.
        '403':
          description: An action names a queue or endpoint outside the grant (`forbidden`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '404':
          description: No such webhook in the caller's reach.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '409':
          description: The webhook's endpoint is revoked (`state` names it), or the rules changed while the save was being checked (`conflict`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '413':
          description: The body exceeds 256 KiB.
        '429':
          description: The caller's per-human write budget is spent; see Retry-After.
        '503':
          description: The rule evaluator is unavailable, so the save could not be checked (`unavailable`); retry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
  /api/v1/webhooks/{webhook_id}/rules/test:
    parameters:
      - name: webhook_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [operator-api]
      operationId: testWebhookRules
      summary: Dry-run routing against a stored delivery or a sample payload
      description: |
        `test_webhook_rules`: evaluates the saved rules, or candidate `rules`/`default_action`/
        `params`, against one of THIS webhook's stored deliveries (`event_id`) or a sample `payload`,
        through the evaluator the receiver uses. Saves nothing. Candidates are validated like a save.
      security:
        - operatorOAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WebhookRulesTestIn' }
      responses:
        '200':
          description: Where the delivery would go, the trace, and (unless omitted) the envelope the rules saw.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookRulesTestOut' }
        '400':
          description: Not exactly one of event_id or payload, or an invalid candidate.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '401':
          description: Missing, expired, revoked, or non-operator bearer.
        '403':
          description: A candidate action names a queue or endpoint outside the grant.
        '404':
          description: No such webhook in the caller's reach, or the event is not one of this webhook's deliveries.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
  /api/v1/webhooks/{webhook_id}/routes:
    parameters:
      - name: webhook_id
        in: path
        required: true
        description: A webhook whose owning endpoint belongs to the caller (see listWebhooks).
        schema: { type: string, format: uuid }
    get:
      tags: [operator-api]
      operationId: listWebhookRoutes
      summary: List the endpoints a webhook delivers to
      description: |
        `list_webhook_routes` (ADR-0022), from the same code: the owning endpoint, which is always a
        target, and the explicit routes, newest grant first. A route whose target was revoked or
        whose friendship lapsed is still listed so it can be removed; deliveries skip it.
      security:
        - operatorOAuth: []
      responses:
        '200':
          description: The webhook's fan-out set.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookRoutes' }
        '401':
          description: Missing, expired, revoked, or non-operator bearer.
        '404':
          description: No such webhook in the caller's reach (another human's and an unknown id are the same answer).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
    post:
      tags: [operator-api]
      operationId: addWebhookRouteFromBody
      summary: Add a route, naming the target in the body
      description: |
        The same operation as addWebhookRoute, for a client that holds the target id in a body as
        `add_webhook_route`'s arguments do.
      security:
        - operatorOAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WebhookRouteIn' }
      responses:
        '200':
          description: The route exists (idempotent).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookRouteAdded' }
        '400':
          description: The body does not decode, or has no target_endpoint_id (`invalid_argument`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '401':
          description: Missing, expired, revoked, or non-operator bearer.
        '403':
          description: The target is not routable (`forbidden`); see addWebhookRoute.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '404':
          description: No such webhook in the caller's reach.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '409':
          description: The webhook's endpoint is not active (`state` names it).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '429':
          description: The caller's per-human write budget is spent; see Retry-After.
  /api/v1/webhooks/{webhook_id}/routes/{endpoint_id}:
    parameters:
      - name: webhook_id
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: endpoint_id
        in: path
        required: true
        description: The target endpoint's id (slugs are resolved by the CLI, not here).
        schema: { type: string, format: uuid }
    put:
      tags: [operator-api]
      operationId: addWebhookRoute
      summary: Make an endpoint a delivery target of a webhook
      description: |
        `add_webhook_route`, through the same checks: the target must be an active endpoint of the
        caller, or of a human the caller may deliver to across an approved friend edge (ADR-0010).
        Adding a route that exists changes nothing. Once routed, the endpoint is in the webhook's
        `grant.endpoints`, so its rules may name it.
      security:
        - operatorOAuth: []
      responses:
        '200':
          description: The route exists (idempotent).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookRouteAdded' }
        '401':
          description: Missing, expired, revoked, or non-operator bearer.
        '403':
          description: |
            The target is not routable (`forbidden`): unknown, malformed, revoked, or another human's
            without an approved edge. Every reason is the same body.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '404':
          description: No such webhook in the caller's reach.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '409':
          description: The webhook's endpoint is not active (`state` names it).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '429':
          description: The caller's per-human write budget is spent; see Retry-After.
    delete:
      tags: [operator-api]
      operationId: removeWebhookRoute
      summary: Stop delivering a webhook to an endpoint
      description: |
        `remove_webhook_route`. Succeeds whether or not the route existed; the owning endpoint
        always stays a target. Works on a revoked endpoint's webhook too, since its routes still
        deliver.
      security:
        - operatorOAuth: []
      responses:
        '200':
          description: The route is gone (idempotent).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookRouteRemoved' }
        '401':
          description: Missing, expired, revoked, or non-operator bearer.
        '404':
          description: No such webhook in the caller's reach.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIError' }
        '429':
          description: The caller's per-human write budget is spent; see Retry-After.
...
components:
...
  schemas:
...
    APIError:
      type: object
      description: The human API's error shape (SPEC-0035). `code` is the code the matching MCP verb returns.
      required: [error, code]
      properties:
        error: { type: string, description: 'Human-readable; names the offending rule where there is one.' }
        code: { type: string, example: not_found }
        state: { type: string, description: 'On a 409 for a revoked endpoint: its state.', example: revoked }
    WebhookListOut:
      type: object
      required: [webhook_id, endpoint_id, endpoint_slug, endpoint_state, agent_name, source_type, target_queue, trust_mode, rule_count, has_default_action, has_params, created_at]
      properties:
        webhook_id: { type: string, format: uuid }
        endpoint_id: { type: string, format: uuid }
        endpoint_slug: { type: string, example: forge-router-ab12cd34 }
        endpoint_state: { type: string, enum: [active, revoked] }
        agent_name: { type: string }
        source_type: { type: string, example: gitea }
        target_queue: { type: string, example: triage }
        trust_mode: { type: string, enum: [signed, token] }
        rule_count: { type: integer }
        has_default_action: { type: boolean }
        has_params: { type: boolean }
        created_at: { type: string, format: date-time }
        rotated_at: { type: string, format: date-time }
...
    WebhookRules:
      type: object
      required: [webhook_id, source_type, target_queue, rules, grant]
      properties:
        webhook_id: { type: string, format: uuid }
        source_type: { type: string }
        target_queue: { type: string }
        default_action: { $ref: '#/components/schemas/RuleAction' }
        rules: { type: array, items: { $ref: '#/components/schemas/Rule' } }
        params: { type: object, additionalProperties: true, description: 'Bound as $params in every rule.' }
        grant:
          type: object
          properties:
            queues: { type: array, items: { type: string } }
            endpoints: { type: array, items: { type: string } }
...
    WebhookRulesIn:
      type: object
      required: [rules]
      description: A whole configuration. A GET response is accepted as it stands (read-only fields are ignored).
      properties:
        rules: { type: array, items: { $ref: '#/components/schemas/Rule' }, description: 'The complete ordered list; [] removes every rule.' }
        default_action: { $ref: '#/components/schemas/RuleAction' }
        params:
          type: [object, "null"]
          additionalProperties: true
          description: Omit to keep the stored params; null clears them.
    WebhookRulesTestIn:
      type: object
      description: Exactly one of event_id or payload.
      properties:
        event_id: { type: integer }
        payload: { description: Any JSON, routed as if it had just arrived and verified. }
        headers: { type: object, additionalProperties: { type: string } }
        rules: { type: array, items: { $ref: '#/components/schemas/Rule' } }
        default_action: { $ref: '#/components/schemas/RuleAction' }
        params: { type: object, additionalProperties: true }
        omit_envelope: { type: boolean }
    WebhookRulesTestOut:
      type: object
      required: [decision, trace]
      properties:
        decision:
          type: object
          properties:
            drop: { type: boolean }
            queue: { type: string }
            endpoints: { type: array, items: { type: string } }
            disposition: { type: string, enum: [routed, dropped, faulted] }
            faulted: { type: boolean, description: 'Blocking: the delivery would be recorded and routed nowhere.' }
            unavailable: { type: boolean }
            fault: { type: object, additionalProperties: true }
        trace: { type: object, additionalProperties: true }
        once_key: { type: string }
        work_order: { type: object, additionalProperties: true }
        envelope: { type: object, additionalProperties: true }
  securitySchemes:
    # Documented for clarity only — these are per-provider signature headers, verified by the app
    # (ADR-0003), NOT bearer credentials the client obtains. UI/SSE endpoints have no in-app auth
    # by design (localhost + Caddy forward_auth). See the info.description security posture.
    githubSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: HMAC-SHA256 of the raw body, `sha256=` prefixed.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: '`t=` timestamp + `v1=` HMAC-SHA256 over "{t}.{body}".'
    slackSignature:
      type: apiKey
      in: header
      name: X-Slack-Signature
      description: '`v0=` HMAC-SHA256 over "v0:{ts}:{body}"; paired with X-Slack-Request-Timestamp.'
    operatorOAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        An OPERATOR OAuth grant (ADR-0019/ADR-0023): an opaque access token minted by the
        authorization-code + PKCE flow with resource = `<base>/api`, bound to the signed-in human.
        Endpoint-bound tokens and static sbk_ credentials do NOT resolve here. The `switchboard`
        CLI performs this flow (gh-style) and stores the rotating pair locally.

# No global security requirement: webhook auth is per-provider signature verification (above),
# and UI/SSE/health endpoints are intentionally unauthenticated at the app layer (ADR-0001).
security: []